requirements.txt
python-telegram-bot>=21.9
openai>=1.57.0
apscheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1
